+++ PRect.java
@@ -1,8 +1,8 @@
 
 public class PRect extends PObjct{
-    int x, y;
+    float x, y;
     int width, height;
-    public PRect(int x, int y, int width, int height) {
+    public PRect(float x, float y, int width, int height) {
         this.x = x;
         this.y = y;
         this.width = width;
